pyproject.toml
[tool.poetry]
name = "grpc_sample"
version = "0.1.0"
description = ""
authors = ["sasano8 <y-sasahara@ys-method.com>"]

[tool.poetry.dependencies]
python = "^3.8"

[tool.poetry.dev-dependencies]
grpcio-tools = "^1.45.0"
grpcio = "^1.45.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
